refactor: extract fixtures and group tests by phase in test_receive_report_case_tree.py
Closes #1090

- Move 10 fixtures from test_receive_report_case_tree.py to
  test/core/behaviors/case/conftest.py so sibling tree test files can
  share them (datalayer, actor_id, reporter_actor_id, actor,
  reporter_actor, report, reporter_accepted_status,
  vendor_received_status, offer, bridge)
- Group the 21 flat test functions into 5 classes by tree phase:
  TestTreeStructure, TestTreeFlow, TestTreeIdempotency,
  TestParticipantCreation, TestEmbargoInitialization
- Move local imports (CVDRole, EM, PEC, CheckCaseExistsForReport, etc.)
  to module level

All 21 tests pass. No logic changes.
